# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
...
    "/rooms/{roomId}/archive/items/{itemId}/comments" : {
      "get" : {
        "tags" : [ "archive-comment-controller" ],
        "operationId" : "list_3",
        "parameters" : [ {
          "name" : "roomId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "itemId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ArchiveCommentResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "archive-comment-controller" ],
        "operationId" : "create_4",
        "parameters" : [ {
          "name" : "roomId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "itemId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateArchiveCommentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArchiveCommentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/rooms/{roomId}/archive/items/{itemId}/comments/{commentId}" : {
      "delete" : {
        "tags" : [ "archive-comment-controller" ],
        "operationId" : "delete_3",
        "parameters" : [ {
          "name" : "roomId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "itemId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      },
      "patch" : {
        "tags" : [ "archive-comment-controller" ],
        "operationId" : "update_4",
        "parameters" : [ {
          "name" : "roomId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "itemId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateArchiveCommentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArchiveCommentResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "MemberBriefResponse" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "nickname" : {
            "type" : "string"
          },
          "profileImage" : {
            "type" : "string"
          }
        }
      },
      "CreateArchiveCommentRequest" : {
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "string"
          }
        }
      },
      "ArchiveCommentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "author" : {
            "$ref" : "#/components/schemas/MemberBriefResponse"
          },
          "body" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "UpdateArchiveCommentRequest" : {
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}